JVCL - renamed method getNumLayers because it was the same as the domain.Map Class method.
There was a small problem with the prototypes for settile and gettile
being x,y,i  when they should be x,y,z to avoid future confusion - I
modified them.

Added a method to TileGroup that enables a tile group to be drawn
exclusively to a single map layer & override each tiles given layer
assignment.